+++ CHANGELOG.md
@@ -3,5 +3,10 @@
 # changed-tests-runner Changelog
 
 ## [Unreleased]
-### Added
-- Initial scaffold created from [IntelliJ Platform Plugin Template](https://github.com/JetBrains/intellij-platform-plugin-template)
+### Initial release
+- Running JUnit >=4 tests based on uncommitted changes from your VCS.
+- Displaying results in a dedicated tool window.
+- Linking executed test suites to the source code.
+- Saving run configurations in the project settings.
+- Persisting plugin settings in the IDE settings.
+- Action is available from a dedicated tool window or Tools menu.